Exposes mixins for use with other element base classes
